requirements.txt
mysql-connector-python==9.1.0 
watchdog==5.0.3